Restructure tests, and nil check
